Logging and exception improvements and cleanups
Added a function to retreive a stack backtrace on systems that support it
(requires glibc and linking with -rdynamic).

Extended exception class to contain a field for stack backtrace. Exception
printing redirected to new MLOG facility.

Added new logging macros MLOG ("message") and DLOG ("debug") that
pretty-print, prefix (with time and function name) and correctly print
multi-line debug messages. Log-style line prefixing is active only if
LOGGING environment variable is set.

MLOG is meant for general-purpose application logging. DLOG is meant for
debugging, and should be safe to turn off (in the future) in production
code.