Config class: Added "inline variable overrides"
The name of the file passed to Config::load() can now have an optional
appendage in the form {var1=val1;var2=val2;...}, where the contents will be
appended to the filename and will override any variables of the same name
that are in the file.

Note: any variable expansion happens after the overrides have been loaded.